+++ public/src/js/yap-core.js
@@ -1071,6 +1071,28 @@ function openServiceBodyCallHandling(service_body_id) {
     });
 }
 
+function deleteVoicemail(callsid) {
+    spinnerDialog(true, "Marking voicemail as deleted...", function () {
+        $.ajax({
+            async: false,
+            type: "POST",
+            url: "../v1/events/status",
+            data: {
+                "callsid": callsid,
+                "status": 1,
+                "event_id": 4 // VOICEMAIL
+            },
+            complete: function() {
+                spinnerDialog(false);
+                location.reload();
+            },
+            timeout: 60000
+        });
+    });
+
+    return false;
+}
+
 function groupsPage() {
     $("#group_id").on("change", function() {
         addNewVolunteerDialog($(this).val() >= 0);