+++ routes/web.php
@@ -15,6 +15,9 @@ use Illuminate\Support\Facades\Route;
 //Route::get("/storage/{extra}", function ($extra) {
 //    return redirect("/public/storage/$extra");
 //})->where("extra", ".*");
+Route::get('/v1/events/status', 'App\Http\Controllers\EventStatusController@index');
+Route::post('/v1/events/status', 'App\Http\Controllers\EventStatusController@set');
+Route::get("/admin/auth/rights", 'App\Http\Controllers\AuthController@rights');
 Route::get("/admin/auth/logout", 'App\Http\Controllers\AuthController@logout');
 Route::get("/admin/auth/timeout", 'App\Http\Controllers\AuthController@timeout');
 Route::get("/admin/auth/invalid", 'App\Http\Controllers\AuthController@invalid');